Characterize targets using additional resources

Append target characteristics to the targets table in the associatoR object. Use pre-defined characteristics from psycho-linguistic research or supply a data.frame as a user-defined look-up table.

Usage

ar_characterize_targets(
  associations,
  characteristics = c("valence", "arousal", "dominance", "concreteness",
    "word_frequency"),
  case_sensitive = FALSE
)

Arguments

associations

an associatoR object containing association data as generated by ar_import.

characteristics

a character string or vector specifying the characteristic(s) containing one or more from c("valence", "arousal", "dominance", "concreteness", "word_frequency") or a data.frame containing characteristics. The data.frame must contain a column word to serve as the look up key and may contain additional columns with word characteristics.

case_sensitive

a logical specifying whether to the case of targets. Default is FALSE.

Value

Returns an associatoR object containing a list of tibbles, with targets gaining additional columns including the target characteristics:

participants

A tibble of participants including a participant id and potential participant attributes.

cues

A tibble of cues including a cue variable and potential cue attributes.

responses

A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.

targets

A tibble of targets including the specified analysis targets, and word characteristics

Details

Data on valence, arousal, and dominance are retrieved from Warriner et al. (2013). Concreteness and word frequency data are retrieved from Brysbaert et al. (2014).

References

Warriner, A. B., Kuperman, V., & Brysbaert, M. (2013). Norms of valence, arousal, and dominance for 13,915 English lemmas. Behavior Research Methods, 45(4), 1191–1207. https://doi.org/10.3758/s13428-012-0314-x

Brysbaert, M., Warriner, A. B., & Kuperman, V. (2014). Concreteness ratings for 40 thousand generally known English word lemmas. Behavior Research Methods, 46(3), 904–911. https://doi.org/10.3758/s13428-013-0403-5

Examples

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_set_targets(targets = "cues") %>%
  ar_characterize_targets()
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <dbl> <chr>  <chr>      
#> 1     1 male   high school
#> 2     2 male   high school
#> 3     3 male   high school
#> 4     4 male   high school
#> 5     5 male   high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 804 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 Einstein    
#> 3 books       
#> 4 IQ tests    
#> 5 college     
#> # ℹ 799 more rows
#> 
#> responses
#> # A tibble: 29,882 × 5
#>      id cue          response     response_position response_level
#>   <dbl> <chr>        <chr>                    <dbl>          <dbl>
#> 1     1 intelligence Einstein                     1              1
#> 2     1 intelligence books                        2              1
#> 3     1 intelligence IQ tests                     3              1
#> 4     1 intelligence college                      4              1
#> 5     1 intelligence smart people                 5              1
#> # ℹ 29,877 more rows
#> 
#> targets
#> # A tibble: 804 × 6
#>   target       valence arousal dominance concreteness word_frequency
#>   <chr>          <dbl>   <dbl>     <dbl>        <dbl>          <dbl>
#> 1 intelligence    7.65    6.32      6.72         2.24            983
#> 2 Einstein       NA      NA        NA           NA                NA
#> 3 books          NA      NA        NA           NA                NA
#> 4 IQ tests       NA      NA        NA           NA                NA
#> 5 college         6.44    4         5.89         4.62           4344
#> # ℹ 799 more rows
#> 

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_set_targets(targets = "cues") %>%
  ar_characterize_targets(characteristics = data.frame(word = c("intelligence", "brain"),
                                                       message = c("Hello", "World")))
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <dbl> <chr>  <chr>      
#> 1     1 male   high school
#> 2     2 male   high school
#> 3     3 male   high school
#> 4     4 male   high school
#> 5     5 male   high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 804 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 Einstein    
#> 3 books       
#> 4 IQ tests    
#> 5 college     
#> # ℹ 799 more rows
#> 
#> responses
#> # A tibble: 29,882 × 5
#>      id cue          response     response_position response_level
#>   <dbl> <chr>        <chr>                    <dbl>          <dbl>
#> 1     1 intelligence Einstein                     1              1
#> 2     1 intelligence books                        2              1
#> 3     1 intelligence IQ tests                     3              1
#> 4     1 intelligence college                      4              1
#> 5     1 intelligence smart people                 5              1
#> # ℹ 29,877 more rows
#> 
#> targets
#> # A tibble: 804 × 2
#>   target       message
#>   <chr>        <chr>  
#> 1 intelligence Hello  
#> 2 Einstein     NA     
#> 3 books        NA     
#> 4 IQ tests     NA     
#> 5 college      NA     
#> # ℹ 799 more rows
#>