Cluster targets

ar_cluster_targets produces a target clustering based on the associatoR object's target embedding.

Usage

ar_cluster_targets(
  associations,
  method = "louvain",
  similarity = "arccos",
  k = NULL,
  linkage = NULL,
  eps = NULL,
  ...
)

Arguments

associations

an associatoR object including target_embeddings.

method

a character specifying the clustering method. One of c("louvain", "hclust", "kmeans", "dbscan"). Default is "louvain".

similarity

a character specifying the similarity metric. One of c("arccos", "cosine", "euclidean"). Default is "arccos".

k

an integer specifying the number of clusters for method = c("hclust", "kmeans").

linkage

a character specifying the linkage criterion for method = c("hclust"). Default is "complete".

eps

a numeric specifying the point distance used in method = c("dbscan").

...

additional attributes passed on to the clustering method.

Value

The function returns the associatoR object including a new

clusters column in targets containing the target clusters.

Examples

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_set_targets(targets = "cues") %>%
  ar_embed_targets() %>%
  ar_cluster_targets(method = "louvain")
#> Warning: 432 targets with count < min_count were dropped from embedding.
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <dbl> <chr>  <chr>      
#> 1     1 male   high school
#> 2     2 male   high school
#> 3     3 male   high school
#> 4     4 male   high school
#> 5     5 male   high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 792 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 Einstein    
#> 3 books       
#> 4 science     
#> 5 technology  
#> # ℹ 787 more rows
#> 
#> responses
#> # A tibble: 29,892 × 5
#>      id cue          response   response_position response_level
#>   <dbl> <chr>        <chr>                  <dbl>          <dbl>
#> 1     1 intelligence Einstein                   1              1
#> 2     1 intelligence books                      2              1
#> 3     1 intelligence science                    3              1
#> 4     1 intelligence technology                 4              1
#> 5     1 intelligence university                 5              1
#> # ℹ 29,887 more rows
#> 
#> targets
#> # A tibble: 792 × 2
#>   target       cluster  
#>   <chr>        <chr>    
#> 1 intelligence cluster_1
#> 2 Einstein     cluster_2
#> 3 books        cluster_3
#> 4 science      cluster_2
#> 5 technology   cluster_2
#> # ℹ 787 more rows
#> 
#> target_embedding
#> # A tibble: 360 × 301
#>   target       dim_1   dim_2  dim_3  dim_4  dim_5  dim_6   dim_7 dim_8
#>   <chr>        <dbl>   <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl> <dbl>
#> 1 intelligence  4.18 -0.153  -1.50  -0.628  1.15   -3.88   2.24  -1.51
#> 2 Einstein      3.18  0.0482  0.247  2.73   2.87   -4.19  -0.981 -1.97
#> 3 books         9.60 -1.84    9.14  -0.268 16.1     7.52   8.88   9.98
#> 4 science       9.40  8.58    0.929  5.69   9.32  -12.3  -16.0    3.21
#> 5 technology    6.42 27.2    -1.95  -1.75  -0.880   6.04   2.98  -4.49
#> # ℹ 355 more rows
#> # ℹ 292 more variables: dim_9 <dbl>, dim_10 <dbl>, dim_11 <dbl>,
#> #   dim_12 <dbl>, dim_13 <dbl>, dim_14 <dbl>, dim_15 <dbl>,
#> #   dim_16 <dbl>, dim_17 <dbl>, dim_18 <dbl>, dim_19 <dbl>,
#> #   dim_20 <dbl>, dim_21 <dbl>, dim_22 <dbl>, dim_23 <dbl>,
#> #   dim_24 <dbl>, dim_25 <dbl>, dim_26 <dbl>, dim_27 <dbl>,
#> #   dim_28 <dbl>, dim_29 <dbl>, dim_30 <dbl>, dim_31 <dbl>, …
#>