Count target occurrences in responses

Append the response frequency of each target to the targets table in the associatoR object.

Usage

ar_count_targets(associations, ...)

Arguments

associations

an associatoR object containing association data as generated by ar_import.

...

optional logical comparisons filtering the responses before computing the frequencies.

Value

Returns an associatoR object containing a list of tibbles, with targets gaining a frequency column:

participants

A tibble of participants including a participant id and potential participant attributes.

cues

A tibble of cues including a cue variable and potential cue attributes.

responses

A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.

targets

A tibble of targets including a column `frequency` containing counts of this target in the responses.

Examples

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_set_targets(targets = "cues") %>%
  ar_count_targets()
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <int> <chr>  <chr>      
#> 1     1 female high school
#> 2     2 female high school
#> 3     3 female high school
#> 4     4 female high school
#> 5     5 female high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 553 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 Brain       
#> 3 Knowledge   
#> 4 Wisdom      
#> 5 IQ          
#> # ℹ 548 more rows
#> 
#> responses
#> # A tibble: 29,820 × 5
#>      id cue          response  response_position response_level
#>   <int> <chr>        <chr>                 <int>          <dbl>
#> 1     1 intelligence Brain                     1              1
#> 2     1 intelligence Knowledge                 2              1
#> 3     1 intelligence Wisdom                    3              1
#> 4     1 intelligence IQ                        4              1
#> 5     1 intelligence Learning                  5              1
#> # ℹ 29,815 more rows
#> 
#> targets
#> # A tibble: 553 × 2
#>   target       frequency
#>   <chr>            <dbl>
#> 1 intelligence       106
#> 2 Brain              131
#> 3 Knowledge          235
#> 4 Wisdom             157
#> 5 IQ                 136
#> # ℹ 548 more rows
#> 

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_set_targets(targets = "cues") %>%
  ar_count_targets(response_position == 1)
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <int> <chr>  <chr>      
#> 1     1 female high school
#> 2     2 female high school
#> 3     3 female high school
#> 4     4 female high school
#> 5     5 female high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 553 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 Brain       
#> 3 Knowledge   
#> 4 Wisdom      
#> 5 IQ          
#> # ℹ 548 more rows
#> 
#> responses
#> # A tibble: 29,820 × 5
#>      id cue          response  response_position response_level
#>   <int> <chr>        <chr>                 <int>          <dbl>
#> 1     1 intelligence Brain                     1              1
#> 2     1 intelligence Knowledge                 2              1
#> 3     1 intelligence Wisdom                    3              1
#> 4     1 intelligence IQ                        4              1
#> 5     1 intelligence Learning                  5              1
#> # ℹ 29,815 more rows
#> 
#> targets
#> # A tibble: 553 × 2
#>   target       frequency
#>   <chr>            <dbl>
#> 1 intelligence        23
#> 2 Brain               89
#> 3 Knowledge           31
#> 4 Wisdom              27
#> 5 IQ                  63
#> # ℹ 548 more rows
#>