Embed associations

ar_embed generates target embeddings.

Usage

ar_embed(
  associations,
  type = "ppmi-svd",
  min_count = 5,
  n_dim = 300,
  model = NULL,
  token = NULL,
  context = NULL
)

Arguments

associations

an associatoR object including targets.

type

a character specifying the type of embedding. One of c("counts","ppmi","ppmi-svd","huggingface"). Default is "ppmi-svd".

min_count

an integer value specifying the minimum response count for responses to be considered in the embedding for type = c("counts","ppmi","ppmi-svd"). Default is 5.

n_dim

an integer value specifying the number of dimensions generated in type = "ppmi-svd". Default is 300.

model

a character specifying the model label. Must match the name on huggingface.co/models.

token

a character string specifying the access token for the hugging face API. Must be obtained from huggingface.co/inference-api.

context

an optional character string specifying a common lead text that may help the language model interpret the associations. Defaults to "Free association: "

Value

The function returns the associatoR object including a new

matrix element called target_embeddings containing the target embeddings.

References

Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv

Examples

ar_import(risk_asso, participant = id, cue = cue, response = response,
          response_vars = trial, participant_vars = c(gender, age, age_group)) %>%
  ar_normalize() %>%
  ar_set_targets(targets = "responses") %>%
  ar_embed()
#> Warning: 915 cues with < min_count responses >= min_count were dropped from embedding.
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,205 × 4
#>      id gender   age age_group
#>   <dbl> <chr>  <dbl>     <dbl>
#> 1     1 female    33         2
#> 2     2 male      56         4
#> 3     3 female    27         1
#> 4     4 male      66         5
#> 5     5 female    59         5
#> # ℹ 1,200 more rows
#> 
#> cues
#> # A tibble: 1,439 × 1
#>   cue           
#>   <chr>         
#> 1 risiko        
#> 2 wohnortwechsel
#> 3 angst         
#> 4 einschätzbar  
#> 5 wetterchaos   
#> # ℹ 1,434 more rows
#> 
#> responses
#> # A tibble: 36,053 × 4
#>      id cue    response       trial
#>   <dbl> <chr>  <chr>          <dbl>
#> 1     1 risiko wohnortwechsel     1
#> 2     2 risiko angst              1
#> 3     3 risiko einschätzbar       1
#> 4     4 risiko wetterchaos        1
#> 5     5 risiko schwierigkeit      1
#> # ℹ 36,048 more rows
#> 
#> targets
#> # A tibble: 9,138 × 1
#>   target        
#>   <chr>         
#> 1 wohnortwechsel
#> 2 angst         
#> 3 einschätzbar  
#> 4 wetterchaos   
#> 5 schwierigkeit 
#> # ℹ 9,133 more rows
#> 
#> target_embedding
#> # A tibble: 518 × 301
#>   target       dim_1  dim_2  dim_3   dim_4   dim_5 dim_6  dim_7  dim_8
#>   <chr>        <dbl>  <dbl>  <dbl>   <dbl>   <dbl> <dbl>  <dbl>  <dbl>
#> 1 abenteuer   -8.01   4.47   4.57  -3.72   -7.34   3.54   1.98  -0.354
#> 2 abenteuerl… -0.810  0.577 -0.728 -0.0210 -1.67   1.28   1.33   1.62 
#> 3 abenteuerl… -2.28   0.568  1.32  -0.574  -2.34   0.394  0.500  0.515
#> 4 abenteueru… -0.914  0.828  0.408  0.165  -1.45   0.640  0.543 -0.901
#> 5 abgabe      -1.33  -0.839 -0.541 -0.376   0.0865 1.03  -1.04  -0.151
#> # ℹ 513 more rows
#> # ℹ 292 more variables: dim_9 <dbl>, dim_10 <dbl>, dim_11 <dbl>,
#> #   dim_12 <dbl>, dim_13 <dbl>, dim_14 <dbl>, dim_15 <dbl>,
#> #   dim_16 <dbl>, dim_17 <dbl>, dim_18 <dbl>, dim_19 <dbl>,
#> #   dim_20 <dbl>, dim_21 <dbl>, dim_22 <dbl>, dim_23 <dbl>,
#> #   dim_24 <dbl>, dim_25 <dbl>, dim_26 <dbl>, dim_27 <dbl>,
#> #   dim_28 <dbl>, dim_29 <dbl>, dim_30 <dbl>, dim_31 <dbl>, …
#>