Import association data

ar_import processes tabular association data into an associatoR object.

Usage

ar_import(
  data,
  participant,
  cue,
  response,
  participant_vars = NULL,
  cue_vars = NULL,
  response_vars = NULL
)

Arguments

data

a data.frame or tibble containing the association data.

participant

variable name in data identifying the participants.

cue

variable name in data or character string identifying the cue(s).

response

variable name in data identifying the responses.

participant_vars

optional variable names in data identifying participant attributes (e.g., demographics).

cue_vars

optional variable names in data identifying cue attributes.

response_vars

optional variable names in data identifying response attributes.

Value

Returns an associatoR object containing a list of tibbles:

participants

A tibble of participants including a participant id and potential participant attributes.

cues

A tibble of cues including a cue variable and potential cue attributes.

responses

A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.

References

Aeschbach, S., Mata, R., Wulff, D. U. (in progress)

Examples

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level))
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <dbl> <chr>  <chr>      
#> 1     1 male   high school
#> 2     2 male   high school
#> 3     3 male   high school
#> 4     4 male   high school
#> 5     5 male   high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 804 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 Einstein    
#> 3 books       
#> 4 IQ tests    
#> 5 college     
#> # ℹ 799 more rows
#> 
#> responses
#> # A tibble: 29,882 × 5
#>      id cue          response     response_position response_level
#>   <dbl> <chr>        <chr>                    <dbl>          <dbl>
#> 1     1 intelligence Einstein                     1              1
#> 2     1 intelligence books                        2              1
#> 3     1 intelligence IQ tests                     3              1
#> 4     1 intelligence college                      4              1
#> 5     1 intelligence smart people                 5              1
#> # ℹ 29,877 more rows
#>