Normalize responses manually

ar_normalize_manual offers an interface to add self-defined normalization steps for responses (and cues) in associatoR objects.

ar_normalize_manual offers an interface to add self-defined normalization steps for responses (and cues) in associatoR objects.

Usage

ar_normalize_manual(associations, fun, ..., process_cues = FALSE)

ar_normalize_manual(associations, fun, ..., process_cues = FALSE)

Arguments

associations

an associatoR object containing association data as generated by ar_import.

fun

a vectorized function applied to all responses and cues in the response matrix, excluding cues that exist in the .

...

additional arguments passed to fun.

process_cues

a logical indicating if cues should be processed (i.e. changed by fun) or not.

dict

a data.frame consisting of two columns named c("old","new").

Value

Returns an associatoR object containing a list of tibbles:

participants

A tibble of participants including a participant id and potential participant attributes.

cues

A tibble of cues including a cue variable and potential cue attributes.

responses

A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes. All responses are passed through fun

Returns an associatoR object containing a list of tibbles:

participants

A tibble of participants including a participant id and potential participant attributes.

cues

A tibble of cues including a cue variable and potential cue attributes.

responses

A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes. All responses are passed through fun

References

Aeschbach, S., Mata, R., Wulff, D. U. (in progress)

Aeschbach, S., Mata, R., Wulff, D. U. (in progress)

Examples

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_normalize_manual(trimws, which = "left")
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <int> <chr>  <chr>      
#> 1     1 female high school
#> 2     2 female high school
#> 3     3 female high school
#> 4     4 female high school
#> 5     5 female high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 553 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 Brain       
#> 3 Knowledge   
#> 4 Wisdom      
#> 5 IQ          
#> # ℹ 548 more rows
#> 
#> responses
#> # A tibble: 29,820 × 6
#>      id cue          response  response_position response_level response_orig
#>   <int> <chr>        <chr>                 <int>          <dbl> <chr>        
#> 1     1 intelligence Brain                     1              1 Brain        
#> 2     1 intelligence Knowledge                 2              1 Knowledge    
#> 3     1 intelligence Wisdom                    3              1 Wisdom       
#> 4     1 intelligence IQ                        4              1 IQ           
#> 5     1 intelligence Learning                  5              1 Learning     
#> # ℹ 29,815 more rows
#> 

dict = tibble(old = c("intelligence", "iq"),
              new = c("Intelligence", "IQ"))
#> Error in tibble(old = c("intelligence", "iq"), new = c("Intelligence",     "IQ")): could not find function "tibble"

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_normalize_dict(dict)
#> Error in ar_normalize_dict(., dict): could not find function "ar_normalize_dict"