Subset associations

Subset the associatoR object.

Usage

ar_subset(associations, ...)

Arguments

associations

an associatoR object containing association data as generated by ar_import.

...

optional logical comparisons filtering the responses before computing the frequencies.

Value

Returns an associatoR object containing a list of tibbles:

participants

A tibble of participants including a participant id and potential participant attributes.

cues

A tibble of cues including a cue variable and potential cue attributes.

responses

A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.

Details

Note that subsets resets the associatoR object. A potential target tibble will be removed and has to be added again using ar_set_target.

Examples

ar_import(fa_data,
          participant = "participantID", participant_vars = c("age", "gender"),
          cue = "cue", response = "response",
          response_vars = c("created_at", "pos")) %>%
  ar_subset(age > 25, gender == "Fe")
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 496 × 3
#>      id   age gender
#>   <dbl> <dbl> <chr> 
#> 1    77    34 Fe    
#> 2   166    29 Fe    
#> 3   513    27 Fe    
#> 4   788    28 Fe    
#> 5   897    54 Fe    
#> # ℹ 491 more rows
#> 
#> cues
#> # A tibble: 1 × 1
#>   cue  
#>   <chr>
#> 1 cue  
#> 
#> responses
#> # A tibble: 1,494 × 5
#>      id cue   response created_at          pos  
#>   <dbl> <chr> <chr>    <dttm>              <chr>
#> 1    77 cue   he       2011-08-15 04:15:38 R1   
#> 2    77 cue   is       2011-08-15 04:15:38 R2   
#> 3    77 cue   will     2011-08-15 04:15:38 R3   
#> 4   166 cue   her      2011-08-15 12:46:08 R1   
#> 5   166 cue   woman    2011-08-15 12:46:08 R2   
#> # ℹ 1,489 more rows
#>