Export wordlist

ar_wordlist_export() exports a table of responses for manual spelling correction.

Usage

ar_wordlist_export(
  associations,
  file = "wordlist_correction.csv",
  check_spelling = TRUE,
  check_acronym = TRUE
)

Arguments

associations

an associatoR object containing association data as generated by ar_import.

file

a character string specifying the filename including folders. Default is "wordlist_correction.csv".

check_spelling

a logical specifying whether to include a column identifying likely mispelled words based on the spell_check_text. Note that currently the spell check only works for English associations. Default is TRUE.

check_acronym

a character specifying whether to include columns identifying potential acronyms. Acronyms are identified in two first. First, responses are checked against a Wikipedia list ("https://en.wikipedia.org/wiki/Lists_of_acronyms"). Matching of the original and lower-case responses triggers "likely an acronym" and "potentially an acronym". The matched acronym is spelled out in an additional column named acronym_candidate. Second, the a simple acronym detection based on letter casing also triggers "likely an acronym" but does not produce values in acronym_candidate. Default is TRUE.

Value

Writes a .CSV file containing a list of unique responses plus additional information.

References

Aeschbach, S., Mata, R., Wulff, D. U. (in preparation)

Examples

ar_obj = ar_import(intelligence,
                   participant = participant_id,
                   cue = cue,
                   response = response,
                   participant_vars = c(gender, education),
                   response_vars = c(response_position, response_level)) %>%
  ar_normalize_manual(trimws, which = "left")

ar_wordlist_export(ar_obj)