Write a csv file

ar_write_csv() joins all data in an associatoR object (except potential target embedding) into one table and writes the table to disk using write_csv.

Usage

ar_write_csv(associations, file, ...)

Arguments

associations

an associatoR object containing association data as generated by ar_import.

file

a character to write to.

...

optional arguments passed to readr::write_csv()

Value

ar_write_csv() returns the joined associations as a tibble invisibly.

Examples

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_set_targets(targets = "cues") %>%
  ar_write_csv(file = "intelligence.csv")