Characterize targets using external lookup-table

Append target characteristic to the targets table in the associatoR object

Usage

ar_characterize_targets(
  associations,
  characteristic = c("valence", "arousal", "dominance", "concreteness", "frequency")
)

Arguments

associations

an associatoR object containing association data as generated by ar_import.

characteristic

a character string or vector specifying which word characteristics to add to targets. Must be one or multiple of c("valence", "arousal", "dominance", "concreteness", "frequency"). Data on valence, arousal, and dominance are retrieved from Warriner et al. (2013), concreteness and frequency data are retrieved from Brysbaert et al. (2014).

Value

Returns an associatoR object containing a list of tibbles, with targets gaining additional columns including the target characteristics:

participants

A tibble of participants including a participant id and potential participant attributes.

cues

A tibble of cues including a cue variable and potential cue attributes.

responses

A tibble of responses including a participant id, the cues, the responses, the response level, and additional response attributes.

targets

A tibble of targets including the specified analysis target, and word characteristics

References

Warriner, A.B., Kuperman, V. & Brysbaert, M. Norms of valence, arousal, and dominance for 13,915 English lemmas. Behav Res 45, 1191–1207 (2013). https://doi.org/10.3758/s13428-012-0314-x

Brysbaert, M., Warriner, A.B. & Kuperman, V. Concreteness ratings for 40 thousand generally known English word lemmas. Behav Res 46, 904–911 (2014). https://doi.org/10.3758/s13428-013-0403-5

Examples

ar_import(fa_data,
          participant = "participantID", participant_vars = c("age", "gender"),
          cue = "cue", response = "response",
          response_vars = c("created_at", "pos")) %>%
  ar_normalize() %>%
  ar_set_targets("cues") %>%
  ar_characterize_targets()
#> New names:
#> • `` -> `...1`
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 997 × 3
#>      id   age gender
#>   <dbl> <dbl> <chr> 
#> 1    77    34 Fe    
#> 2   166    29 Fe    
#> 3   223    31 Ma    
#> 4   513    27 Fe    
#> 5   577    20 Fe    
#> # ℹ 992 more rows
#> 
#> cues
#> # A tibble: 10 × 1
#>   cue       
#>   <chr>     
#> 1 she       
#> 2 capital   
#> 3 paper clip
#> 4 recess    
#> 5 neon      
#> # ℹ 5 more rows
#> 
#> responses
#> # A tibble: 3,000 × 5
#>      id cue   response created_at          pos  
#>   <dbl> <chr> <chr>    <dttm>              <chr>
#> 1    77 she   he       2011-08-15 04:15:38 R1   
#> 2    77 she   is       2011-08-15 04:15:38 R2   
#> 3    77 she   will     2011-08-15 04:15:38 R3   
#> 4   166 she   her      2011-08-15 12:46:08 R1   
#> 5   166 she   woman    2011-08-15 12:46:08 R2   
#> # ℹ 2,995 more rows
#> 
#> targets
#> # A tibble: 10 × 6
#>   target     valence arousal dominance concreteness frequency
#>   <chr>        <dbl>   <dbl>     <dbl>        <dbl>     <dbl>
#> 1 she          NA      NA        NA            3.36    190377
#> 2 capital       5.29    3.14      5.85         4.1        648
#> 3 paper clip   NA      NA        NA            4.79         0
#> 4 recess        6.57    4.15      5.32         3.07       178
#> 5 neon          6.48    4.56      5.62         4.07       146
#> # ℹ 5 more rows
#>