Embed associations

ar_embed generates target embeddings.

Usage

ar_embed(
  associations,
  type = "ppmi-svd",
  min_count = 5,
  n_dim = 300,
  model = NULL,
  token = NULL,
  context = NULL
)

Arguments

associations

an associatoR object including targets.

type

a character specifying the type of embedding. One of c("counts","ppmi","ppmi-svd","huggingface"). Default is "ppmi-svd".

min_count

an integer value specifying the minimum response count for responses to be considered in the embedding for type = c("counts","ppmi","ppmi-svd"). Default is 5.

n_dim

an integer value specifying the number of dimensions generated in type = "ppmi-svd". Default is 300.

model

a character specifying the model label. Must match the name on huggingface.co/models.

token

a character string specifying the access token for the hugging face API. Must be obtained from huggingface.co/inference-api.

context

an optional character string specifying a common lead text that may help the language model interpret the associations. Defaults to "Free association: "

Value

The function returns the associatoR object including a new

matrix element called target_embeddings containing the target embeddings.

References

Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv

Examples

ar_import(risk_asso, participant = "id", cue = "cue", response = "response",
          response_vars = "trial", participant_vars = c("gender", "age", "age_group")) %>%
  ar_normalize() %>%
  ar_set_targets(target_set = "responses") %>%
  ar_embed()
#> Error in ar_set_targets(., target_set = "responses"): unused argument (target_set = "responses")