Embed associations

ar_embed_targets generates target embeddings.

Usage

ar_embed_targets(
  associations,
  method = "ppmi-svd",
  min_count = 5,
  n_dim = 300,
  model = NULL,
  token = NULL,
  context = NULL
)

Arguments

associations

an associatoR object including targets.

method

a character specifying the type of embedding. One of c("counts","ppmi","ppmi-svd","huggingface"). Default is "ppmi-svd".

min_count

an integer value specifying the minimum response count for responses to be considered in the embedding for method = c("counts","ppmi","ppmi-svd"). Default is 5.

n_dim

an integer value specifying the number of dimensions generated in method = "ppmi-svd". Default is 300.

model

a character specifying the model label. Must match the name on huggingface.co/models.

token

a character string specifying the access token for the hugging face API. Must be obtained from huggingface.co/inference-api.

context

an optional character string specifying a common lead text that may help the language model interpret the associations. Defaults to "Free association: "

Value

The function returns the associatoR object including a new

matrix element called target_embeddings containing the target embeddings.

References

Aeschbach, S., Mata, R., Wulff, D. U. (2024). associatoR. psyArXiv

Examples

ar_import(intelligence,
          participant = participant_id,
          cue = cue,
          response = response,
          participant_vars = c(gender, education),
          response_vars = c(response_position, response_level)) %>%
  ar_set_targets(targets = "cues") %>%
  ar_embed_targets()
#> Warning: 104 targets with count < min_count were dropped from embedding.
#> Warning: all singular values are requested, svd() is used instead
#> 
#> ── An associatoR object ────────────────────────────────────────────────────────
#> 
#> participants
#> # A tibble: 1,000 × 3
#>      id gender education  
#>   <dbl> <chr>  <chr>      
#> 1     1 female high school
#> 2     2 female high school
#> 3     3 female high school
#> 4     4 female high school
#> 5     5 female high school
#> # ℹ 995 more rows
#> 
#> cues
#> # A tibble: 359 × 1
#>   cue         
#>   <chr>       
#> 1 intelligence
#> 2 brain       
#> 3 knowledge   
#> 4 wisdom      
#> 5 iq          
#> # ℹ 354 more rows
#> 
#> responses
#> # A tibble: 29,820 × 5
#>      id cue          response  response_position response_level
#>   <dbl> <chr>        <chr>                 <dbl>          <dbl>
#> 1     1 intelligence brain                     1              1
#> 2     1 intelligence knowledge                 2              1
#> 3     1 intelligence wisdom                    3              1
#> 4     1 intelligence iq                        4              1
#> 5     1 intelligence learning                  5              1
#> # ℹ 29,815 more rows
#> 
#> targets
#> # A tibble: 359 × 1
#>   target      
#>   <chr>       
#> 1 intelligence
#> 2 brain       
#> 3 knowledge   
#> 4 wisdom      
#> 5 iq          
#> # ℹ 354 more rows
#> 
#> target_embedding
#> # A tibble: 255 × 256
#>   target      dim_1  dim_2   dim_3  dim_4  dim_5   dim_6 dim_7   dim_8
#>   <chr>       <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl> <dbl>   <dbl>
#> 1 intelligen… -4.31  0.841 -0.244  -0.253  0.515   0.376  1.51  -0.732
#> 2 brain       -5.58 -4.55  -1.50   -5.54   8.14   14.0    8.68   5.47 
#> 3 knowledge   -6.43 -1.88   1.40    7.17  -2.78    3.22   1.43   0.274
#> 4 wisdom      -6.92 -1.42  -0.0408  3.92   4.47    2.63  -3.88  -0.942
#> 5 iq          -7.21 -2.86   1.64   -2.56   4.54  -10.7   18.8  -10.4  
#> # ℹ 250 more rows
#> # ℹ 247 more variables: dim_9 <dbl>, dim_10 <dbl>, dim_11 <dbl>,
#> #   dim_12 <dbl>, dim_13 <dbl>, dim_14 <dbl>, dim_15 <dbl>,
#> #   dim_16 <dbl>, dim_17 <dbl>, dim_18 <dbl>, dim_19 <dbl>,
#> #   dim_20 <dbl>, dim_21 <dbl>, dim_22 <dbl>, dim_23 <dbl>,
#> #   dim_24 <dbl>, dim_25 <dbl>, dim_26 <dbl>, dim_27 <dbl>,
#> #   dim_28 <dbl>, dim_29 <dbl>, dim_30 <dbl>, dim_31 <dbl>, …
#>